+++ lib/active_record/connection_adapters/sqlserver_adapter.rb
@@ -50,6 +50,11 @@ module ActiveRecord
       # Default precision for 'time' (See https://docs.microsoft.com/en-us/sql/t-sql/data-types/time-transact-sql)
       DEFAULT_TIME_PRECISION = 7
 
+      # FreeTDS-bounded checkout ping timeout (seconds) when TinyTDS provides
+      # Client#ping. Override with config :ping_timeout.
+      # See rails-sqlserver/tiny_tds#609 and activerecord-sqlserver-adapter#1396.
+      DEFAULT_PING_TIMEOUT = 2
+
       attr_reader :spid
 
       cattr_accessor :cs_equality_operator, instance_accessor: false
@@ -143,6 +148,7 @@ module ActiveRecord
         @config[:appname] = self.class.rails_application_name unless @config[:appname]
         @config[:login_timeout] = @config[:login_timeout].present? ? @config[:login_timeout].to_i : nil
         @config[:timeout] = @config[:timeout].present? ? @config[:timeout].to_i / 1000 : nil
+        @config[:ping_timeout] = @config[:ping_timeout].present? ? @config[:ping_timeout].to_i : DEFAULT_PING_TIMEOUT
         @config[:encoding] = @config[:encoding].present? ? @config[:encoding] : nil
 
         @connection_parameters ||= @config
@@ -285,7 +291,14 @@ module ActiveRecord
       # === Abstract Adapter (Connection Management) ================== #
 
       def active?
-        if @raw_connection&.active?
+        return false unless @raw_connection
+
+        if raw_connection_pingable?
+          return false unless @raw_connection.ping(timeout: ping_timeout_seconds)
+
+          verified!
+          true
+        elsif @raw_connection.active?
           verified!
           true
         end
@@ -516,6 +529,16 @@ module ActiveRecord
         @raw_connection_errors ||= [].tap do |errors|
           errors << TinyTds::Error if defined?(TinyTds::Error)
         end
+      end
+
+      def raw_connection_pingable?
+        @raw_connection.respond_to?(:ping)
+      end
+
+      def ping_timeout_seconds
+        seconds = @config[:ping_timeout]
+        seconds = DEFAULT_PING_TIMEOUT if seconds.nil? || seconds.to_i <= 0
+        seconds.to_i
       end
 
       def initialize_dateformatter